refactor multiply sequence to separate file + tests
